Chapter 10 - File Scope Variable Declarations and Storage-Class Specifiers (#11)
* ast: Implement storage class specifiers

* Temporary fixes for ast change

* sema: move to own module

* sema: move passes to their own modules

* sema: update identifier pass for global variables

* sema: Partial implementation of file scope vars

* sema: Ignore file scope vars in non-type passes

* sema: add const eval module

* sema: Disallow storage class idents in for loop decls

* sema: Local externs can never have initalizers

* sema: Redo typechecking pass

Implements global variables and storage class specifiers

* sema: Update initial value for static attributes when declared

* tests: Add test case for static redefinition

* sema: add symbols to `SemaStage`

* sema: make attributes in symbol entries public

* tacky: implement statics and functions as top levels

* codegen: temporary `todo!()`'s to fix compilation

* clippy: fix some warnings

* Little cleaner imo

* Better module design for sema and typechecking pass.

* LSP suggestions.

* Tweaking public re-exports. Making get method public.

* Codegen pass works! Collapsed lots of match statements which were
exclusive to stack offsets to match both stack and data offsets (both
memory addresses) and made a few vars for commonly used registers to
decrease bloat in some of the beefier functions. Asm pass currently just
a todo.

* Sort of working but there is a logic error somewhere causing some of the
test cases to fail.

* Clippy lints.

* Fixed attribute initialization upgrading.

* Fixing typo

* Fixed error where the code produced directly used the external_linkage
of the declaration where a function is defined but the function elided
an earlier declaration which should have resulted in the function being
static.

* Local statics now also get declared in global scope so they get picked
up when allocating space for statics in tacky pass. The same 4 test
cases are still failing, so there must be a logic bug somewhere else.

* Statics behave like statics now :). All test cases pass.

* Fixing static initialization to use zero.

* Moving logic where static declaration is also declared globally (not
that the static should be globally accessible, but we need it in the
global symbol table for tacky stage).

---------

Co-authored-by: kyserd <kyserdy@gmail.com>