tacky: add tacky prefix to temp labels
This is the simplest solution to fix this problem without touching other passes.

The problem is that there exists an edge case where the name of the function and
name of the first argument collide and when positioned as the first parsed thing,
the tacky temporary variable names can collide with the function arguments.

This is the failing test case that caught this:
```c
int a(int a) {
    return a * 2;
}

int main(void) {
    return a(1);
}
```