tacky: don't create new temporaries for initialized variables
I honestly don't know why I wrote this. My current assumption was that when
I was messing with typing tacky tempoararies I accidently clobbered(?) this
line of code. When I came back and realized how wrong it was, I simply fixed
it by fixing the temp creation, not the fact that this never needed a temp at
all.