Chapter 12 - Unsigned Integers (#14)
* lexer: Added support for parsing unsigned and unsigned long integer
suffixes.

* lexer: Fixed hardcoded suffix length offset for unsigned long suffixes.

* ast: Parsing test cases pass now. Currently do not handle truncation
when an integer constant is too large.

* ast: Implementing potentially "correct" behavior when parsing integer
overflow.

* ast: Removing logging print statement.

* ast: Fixing function type formatting.

* typechecking: Making error message styling consistent.

* typechecking: Changing typechecking on functions to perform an equality
check for types (PartialEq is overriden so this should only compare
relevant fields. Got rid of allowance for functions declared without
anything in its parameter list to shadow existing declarations.

* ast: Manual PartialEq implementation since two integers should consider
both the `None` and `Some(true)` cases to be equivalent when checking if
two integers are signed.

* tacky + codegen: Added `ZeroExtend` instruction. Updated parsing to
determine which type of instructions to use in cast expression. Added
`todo!()` match case in codegen so code still compiles.

* codegen + asm: Adding new enum variants and condition codes. Changing
`CondCode` from impl to now also take a signed flag which will be used
to determine which condition code to produce. Currently gets used with a
hardcoded `true` but will need to look up the type from the symbol
table. Adding `todo!()` match case in asm so code still compiles.

* codegen: Adding match case for `MovZeroExtend` instruction.

* codegen + tacky: Made tacky function to get its AST type public. Codegen
now examines tacky value type to determine whether it is signed or
unsigned.

* codegen: Now have signed and unsigned versions of division and modulus.
Implemented helper function to easily create a zero operand for
different operand sizes.

* codegen: Adding cases for `Div` (basically direct copy of `Idiv`).

* codegen: Adding fixup cases for `MovZeroExtend`.

* asm: Adding code emission for `Div` instruction.

* typechecking: TFW LGTM no longer looks good to me.

* typechecking: Resolving initial value for extern declarations to an existing
initial value if there is one.

* tacky: Fixing bug where the target type was being checked when matching
on expression casts rather than the actual value type.

* chore: Removing unnecessary `mut` keyword.

* codegen: Fixing case where a value is being moved with zero extension to
a stack offset. Previously just used r11, but now first moves into r11d
to zero-extend the value then from r11 into destination.

* ast: Copy/paste error 😔

* ast: Fixed bug where unsigned and signed integers of the same size had
the same rank, which caused there to not be implicit casts where they
were needed.

* ast + codegen: Updating match case for when to have a move rather than
just a push (even 32-bit unsigned integers fall out of this range, so
just looking at operand size is not sufficient). Added helper method in
`ast` module to check if a given value can be represented using the
existing constant type.

* codegen + asm: Refactoring bit shifts to correctly use the `shr`
instruction instead of `sar` when working with unsigned integers (does
not need to preserve sign, and should blindly shift bits to the right).

* ast: Fixing comptime integer parsing of potentially very large integers
to use the built in method which converts from a string.

* ast: Fixing overflow case for when there is no suffix added to constant.
Adding a test for this.