tacky + ast + codegen:
tacky: Fixed `Binary` case to correctly cast the results of relational
operators to a boolean rather than using the type of lefthand operand.
This was causing a bunch of `movsd xmm15, 0` instructions to sneak into
the code emission phase which made for a lot of GCC errors.

ast: Added `is_relational` helper method.

codegen: Fixed up a few cases where XMM0 was used but seemed like it
should have been one of the rewrite registers instead.