typechecking: don't special case pointer decay type during conversion
This makes it so that instead of treating pointer decay as a special case, which causes problems when `typecheck_and_convert` is called for a second time on a convered expression, we instead just make it take the address of the first element. This gives array to pointer decay the corre t typeing without any special casing.